chore: Move util functions to logic
