+++ src/logic/nfts/rentals.ts
@@ -1,4 +1,4 @@
-import { NFTCategory, NFTFilters, NFTSortBy } from '@dcl/schemas'
+import { Contract, NFTCategory, NFTFilters, NFTSortBy } from '@dcl/schemas'
 
 export const PROHIBITED_FILTERS: (keyof NFTFilters)[] = [
   'itemRarities',
@@ -47,5 +47,51 @@ export function shouldFetch(filters: NFTFilters): boolean {
       categoriesIsSetAndIsNotLAND ||
       categoriesIsNotSetAndIsNotLand
     ) && Boolean(filters.isOnRent)
+  )
+}
+
+/**
+ * From a list of contracts return the addresses of the land and estate contracts.
+ * Will throw an error if the addresses are not found.
+ * @param contracts The list of contracts from which the land and estate contract addresses will be extracted
+ * @returns An object with both the addresses of the land and estate contracts
+ */
+export function getLandAndEstateContractAddresses(
+  contracts: Contract[]
+): { land: string; estate: string } {
+  let land: string | undefined
+  let estate: string | undefined
+
+  for (const contract of contracts) {
+    if (contract.name === 'LAND') {
+      land = contract.address.toLowerCase()
+    } else if (contract.name === 'Estates') {
+      estate = contract.address.toLowerCase()
+    }
+  }
+
+  if (!land || !estate) {
+    throw new Error('LAND and Estates contracts are required')
+  }
+
+  return {
+    land,
+    estate,
+  }
+}
+
+/**
+ * Util function that determines if the rentals nft component should be used.
+ * @param filters The filters provided in the request which will determine if the rentals nft component should be used.
+ * @returns true if the rentals nft component should be used, false otherwise.
+ */
+export function rentalNFTComponentShouldFetch(filters: NFTFilters): boolean {
+  return (
+    // The lookup for assets on rent is done by another component.
+    !filters.isOnRent &&
+    // Only Lands and Estates can be locked in the rentals contract.
+    !!filters.isLand &&
+    // The rentals subgraph is queried mainly by lessor, se this param is required.
+    !!filters.owner
   )
 }
