+++ src/tests/logic/nft-rentals.spec.ts
@@ -1,7 +1,9 @@
-import { NFTCategory } from '@dcl/schemas'
+import { Contract, NFTCategory, NFTFilters } from '@dcl/schemas'
 import {
+  getLandAndEstateContractAddresses,
   PROHIBITED_FILTERS,
   PROHIBITED_SORTING,
+  rentalNFTComponentShouldFetch,
   shouldFetch,
 } from '../../logic/nfts/rentals'
 
@@ -86,6 +88,104 @@ describe('when checking if the rental should be fetched', () => {
         expect(
           shouldFetch({ isOnRent: true, isLand: true, [filter]: true })
         ).toBe(true)
+      })
+    })
+  })
+})
+
+describe('when calling rentalNFTComponentShouldFetch', () => {
+  let filters: NFTFilters
+
+  beforeEach(() => {
+    filters = {
+      owner: 'owner',
+      isOnRent: false,
+      isLand: true,
+    }
+  })
+
+  describe('when called without owner', () => {
+    beforeEach(() => {
+      delete filters.owner
+    })
+
+    it('should return false', () => {
+      expect(rentalNFTComponentShouldFetch(filters)).toBeFalsy()
+    })
+  })
+
+  describe('when called with isLand as false', () => {
+    beforeEach(() => {
+      filters.isLand = false
+    })
+
+    it('should return false', () => {
+      expect(rentalNFTComponentShouldFetch(filters)).toBeFalsy()
+    })
+  })
+
+  describe('when called with isOnRent as true', () => {
+    beforeEach(() => {
+      filters.isOnRent = true
+    })
+
+    it('should return false', () => {
+      expect(rentalNFTComponentShouldFetch(filters)).toBeFalsy()
+    })
+  })
+
+  describe('when called with the right filters', () => {
+    it('should return true', () => {
+      expect(rentalNFTComponentShouldFetch(filters)).toBeTruthy()
+    })
+  })
+})
+
+describe('when calling getLandAndEstateContractAddresses', () => {
+  let contracts: Contract[]
+
+  beforeEach(() => {
+    contracts = [
+      {
+        name: 'LAND',
+        address: 'LAND_ADDRESS',
+      },
+      {
+        name: 'Estates',
+        address: 'ESTATES_ADDRESS',
+      },
+    ] as Contract[]
+  })
+
+  describe("when the provided contracts don't contain LAND", () => {
+    beforeEach(() => {
+      contracts = [contracts[1]]
+    })
+
+    it('should throw an error indicating a contract is missing', () => {
+      expect(() => getLandAndEstateContractAddresses(contracts)).toThrowError(
+        'LAND and Estates contracts are required'
+      )
+    })
+  })
+
+  describe("when the provided contracts don't contain Estates", () => {
+    beforeEach(() => {
+      contracts = [contracts[0]]
+    })
+
+    it('should throw an error indicating a contract is missing', () => {
+      expect(() => getLandAndEstateContractAddresses(contracts)).toThrowError(
+        'LAND and Estates contracts are required'
+      )
+    })
+  })
+
+  describe('when the provided contracts contain both LAND and Estates', () => {
+    it('should return an object with their addresses lower cased', () => {
+      expect(getLandAndEstateContractAddresses(contracts)).toEqual({
+        land: contracts[0].address.toLowerCase(),
+        estate: contracts[1].address.toLowerCase(),
       })
     })
   })
